+++ README.md
@@ -5,7 +5,7 @@
 
 In IMP-MARL, you can:
 - [Implement your own infrastructure management planning (IMP) environment or execute an available IMP environment](./imp_env/).
-- [Train IMP policies through state-of-the-art MARL methods. The environments can be integrated with typical ecosystems via wrappers](./imp_wrappers/).
+- [Train IMP policies through state-of-the-art MARL methods. The environments can be integrated with typical ecosystems via wrappers](imp_wrapper/).
 - [Compute expert-based heuristic policies](./heuristics/)
 
 Additionally, you will be able to:
@@ -25,7 +25,7 @@ However, to reproduce our results, more packages are required and installation i
 
 ## Tutorials
 - [Create your own IMP environment scenario](imp_env/new_imp_env_tutorial.ipynb)
-- [IMP's API explained](imp_wrappers/wrapper_explained.md)
+- [IMP's API explained](imp_wrapper/wrapper_explained.md)
 - [Train agents like in the paper and/or **reproduce** the results](pymarl/README.md)
 - [Retrieve directly the results](results_scripts/README.md)
