+++ imp_wrapper/pymarl/check_both_env_owf.py
@@ -1,8 +1,8 @@
 """ Test the two PyMarl wrapper with the same owf parameters. """
 
 import numpy as np
-from imp_wrappers.pymarl_ma_struct import PymarlMAStruct
-from imp_wrappers.pymarl_sa_struct import PymarlSAStruct
+from imp_wrappers.pymarl.pymarl_ma_struct import PymarlMAStruct
+from imp_wrappers.pymarl.pymarl_sa_struct import PymarlSAStruct
 
 if __name__ == '__main__':
