+++ imp_wrapper/pymarl/random_pymarl_owf_ma.py
@@ -1,12 +1,11 @@
 """ Demonstration of the PymarlMAStruct wrapper with random actions. """
 
 import numpy as np
-from imp_wrappers.pymarl_ma_struct import PymarlMAStruct
-import os
-if __name__ == '__main__':
+from imp_wrappers.pymarl.pymarl_ma_struct import PymarlMAStruct
 
+if __name__ == '__main__':
     n_episode = 100
-    env = PymarlMAStruct(struct_type="struct",
+    env = PymarlMAStruct(struct_type="owf",
                          n_comp=2,
                          discount_reward=.95,
                          state_obs=True,
@@ -20,7 +19,6 @@ if __name__ == '__main__':
                          campaign_cost=False)
 
     env_info = env.get_env_info()
-
     n_actions = env_info["n_actions"]
     n_agents = env_info["n_agents"]
     array_reward = []
