+++ imp_wrapper/pymarl/random_pymarl_owf_sa.py
@@ -1,11 +1,12 @@
 """ Demonstration of the PymarlMAStruct wrapper with random actions. """
 
 import numpy as np
-from imp_wrappers.pymarl_ma_struct import PymarlMAStruct
-
+from imp_wrappers.pymarl.pymarl_sa_struct import PymarlSAStruct
 if __name__ == '__main__':
+
     n_episode = 100
-    env = PymarlMAStruct(struct_type="owf",
+
+    env = PymarlSAStruct(struct_type="owf",
                          n_comp=2,
                          discount_reward=.95,
                          state_obs=True,
@@ -19,6 +20,7 @@ if __name__ == '__main__':
                          campaign_cost=False)
 
     env_info = env.get_env_info()
+
     n_actions = env_info["n_actions"]
     n_agents = env_info["n_agents"]
     array_reward = []
@@ -33,11 +35,10 @@ if __name__ == '__main__':
             state = env.get_state()
 
             actions = []
-            for k in range(env.n_agents):
-                avail_actions = env.get_avail_agent_actions(k)
-                avail_actions_ind = np.nonzero(avail_actions)[0]
-                action = np.random.choice(avail_actions_ind)
-                actions.append(action)
+            avail_actions = env.get_avail_agent_actions(0)
+            avail_actions_ind = np.nonzero(avail_actions)[0]
+            action = np.random.choice(avail_actions_ind)
+            actions.append(action)
             reward, terminated, info = env.step(actions)
             episode_reward += reward
         array_reward.append(episode_reward)
