+++ imp_wrapper/pymarl/random_pymarl_struct_sa.py
@@ -1,7 +1,7 @@
 """ Demonstration of the PymarlMAStruct wrapper with random actions. """
 
 import numpy as np
-from imp_wrappers.pymarl_sa_struct import PymarlSAStruct
+from imp_wrappers.pymarl.pymarl_sa_struct import PymarlSAStruct
 import os
 
 if __name__ == '__main__':
