--- imp_wrappers/wrapper_explained.md
@@ -1,11 +1,0 @@
-# Wrapper
-
-Explain that wrappers do all the job
-
-You need one wrapper and one imp_env.
-
-Here focus on the PYMARL wrapper and show how we did (link to the notebook).
-
-Example of extention of gym MA and gym SA.
-
-RLLlib simply use gym so I believe we should just have a RLLib how to launch imp_env gym wrapper with gym.
