+++ pymarl/envs/MultiAgentEnv.py
@@ -1,3 +1,4 @@
+""" MultiAgentEnv is the base class for PyMarl environments. """
 class MultiAgentEnv(object):
 
     def step(self, actions):
