+++ pymarl/envs/__init__.py
@@ -1,7 +1,7 @@
 from functools import partial
-from imp_wrappers.MultiAgentEnv import MultiAgentEnv
-from imp_wrappers.pymarl_ma_struct import PymarlMAStruct
-from imp_wrappers.pymarl_sa_struct import PymarlSAStruct
+from pymarl.envs.MultiAgentEnv import MultiAgentEnv
+from imp_wrappers.pymarl.pymarl_ma_struct import PymarlMAStruct
+from imp_wrappers.pymarl.pymarl_sa_struct import PymarlSAStruct
 
 
 def env_fn(env, **kwargs) -> MultiAgentEnv:
